We will have to implement randoms now

+++ src/id0031.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-    int coins[] = { 1, 2, 5, 10, 20, 50, 100, 200 };
+    long long coins[] = { 1, 2, 5, 10, 20, 50, 100, 200 };
     clock_t start = clock();
     long long* partitions = restricted_partition_range(201, coins, 8);
